% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaemixObject.R
\name{coef.saemix}
\alias{coef.saemix}
\alias{coef.SaemixObject}
\alias{coef}
\title{Extract coefficients from a saemix fit}
\usage{
\method{coef}{SaemixObject}(object, ...)
}
\arguments{
\item{object}{a SaemixObject}

\item{...}{further arguments to be passed to or from other methods}
}
\value{
a list with 3 components:
\describe{
\item{fixed}{fixed effects}
\item{population}{a list of population parameters with two elements, a matrix containing the untransformed parameters psi and a matrix containing the transformed parameters phi}
\item{individual}{a list of individual parameters with two elements, a matrix containing the untransformed parameters psi and a matrix containing the transformed parameters phi}
}
}
\description{
Extract coefficients from a saemix fit
}
