% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaemixModel.R
\docType{methods}
\name{plot-SaemixModel}
\alias{plot-SaemixModel}
\alias{plot,SaemixModel-methods}
\title{Plot model predictions using an SaemixModel object}
\arguments{
\item{x}{an SaemixData object or an SaemixSimData object}

\item{y}{unused, present for compatibility with base plot function}

\item{range}{range of X over which the model is to be plotted}

\item{psi}{parameters of the model}

\item{predictors}{additional predictors needed to pass on to the model}

\item{...}{additional arguments to be passed on to plot (titles, legends, ...)}
}
\description{
This function will plot predictions obtained from an SaemixModel object over a given range of X. Additional predictors may be passed on to the function using the predictors argument.
}
\keyword{methods}
