% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_generics.R, R/SaemixObject.R
\docType{methods}
\name{predict-methods}
\alias{predict-methods}
\alias{predict,ANY-method}
\alias{predict,SaemixObject-method}
\title{Methods for Function predict}
\usage{
\S4method{predict}{SaemixObject}(object, newdata = NULL,
  type = c("ipred", "ypred", "ppred", "icpred"), se.fit = FALSE, ...)
}
\arguments{
\item{object}{an SaemixObject}

\item{newdata}{an optional dataframe for which predictions are desired}

\item{type}{the type of predictions (ipred= individual, ypred= mean of the population predictions, ppred=population predictions obtained with the population estimates, icpred=conditional predictions). With newdata, predictions correspond to the population predictions and type is ignored}

\item{se.fit}{whether the SE are to be taken into account in the model predictions}

\item{...}{additional arguments passed on to fitted()}
}
\description{
Methods for function \code{predict}
}
\section{Methods}{
 \describe{

\item{list("signature(object = \"ANY\")")}{Default predict functions}

\item{list("signature(object = \"SaemixObject\")")}{Computes predictions
using the results of an SAEM fit} }
}

\keyword{methods}
