% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_generics.R, R/SaemixData.R,
%   R/SaemixModel.R, R/SaemixRes.R, R/SaemixObject.R
\docType{methods}
\name{print-methods}
\alias{print-methods}
\alias{print.saemix}
\alias{print,ANY-method}
\alias{print,SaemixData-method}
\alias{print,SaemixModel-method}
\alias{print,SaemixRes-method}
\alias{print,SaemixObject-method}
\title{Methods for Function print}
\usage{
\S4method{print}{SaemixData}(x, nlines = 10, ...)

\S4method{print}{SaemixModel}(x, ...)

\S4method{print}{SaemixRes}(x, digits = 2, map = FALSE, ...)

\S4method{print}{SaemixObject}(x, nlines = 10, ...)
}
\arguments{
\item{x}{an object of type SaemixData, SaemixModel, SaemixRes or SaemixObject}

\item{nlines}{maximum number of lines of data to print (defaults to 10)}

\item{...}{additional arguments passed on the print function}

\item{digits}{number of digits to use for pretty printing}

\item{map}{when map is TRUE the individual parameter estimates are shown (defaults to FALSE)}
}
\description{
Prints a summary of an object
}
\section{Methods}{
 \describe{

\item{list("signature(x = \"ANY\")")}{ Default print function }

\item{list("signature(x = \"SaemixData\")")}{ Prints a summary of a
SaemixData object }

\item{list("signature(x = \"SaemixModel\")")}{ Prints a summary of a
SaemixModel object }

\item{list("signature(x = \"SaemixObject\")")}{ Prints a summary of the
results from a SAEMIX fit }

\item{list("signature(x = \"SaemixRes\")")}{ Not user-level } }
}

\keyword{methods}
