% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaemixRes.R, R/SaemixObject.R
\docType{methods}
\name{resid.saemix}
\alias{resid.saemix}
\alias{resid.SaemixRes}
\alias{residuals.SaemixRes}
\alias{resid.SaemixObject}
\title{Extract Model Residuals}
\usage{
\method{resid}{SaemixRes}(object, type = c("ires", "wres", "npde", "pd",
  "iwres", "icwres"), ...)

\method{resid}{SaemixObject}(object, type = c("ires", "wres", "npde",
  "pd", "iwres", "icwres"), ...)
}
\arguments{
\item{object}{an SaemixRes or an SaemixObject object}

\item{type}{string determining which residuals are extracted. Possible values are: "ires" (individual residuals, default), "wres" (weighted population residuals), "npde" (normalised prediction distribution errors), "pd" (prediction discrepancies), "iwres" (individual weighted residuals) and "icwres" (conditional individual weighted residuals). See user guide for details.}

\item{...}{further arguments to be passed to or from other methods}
}
\value{
By default, individual residuals are extracted from the model object
}
\description{
residuals is a generic function which extracts model residuals from objects returned by modelling functions. The  abbreviated form resid is an alias for residuals
}
\keyword{methods}
