% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaemixObject.R
\name{saemix.predict}
\alias{saemix.predict}
\title{Compute model predictions after an saemix fit}
\usage{
saemix.predict(object)
}
\arguments{
\item{object}{an SaemixObject object}
}
\value{
an updated SaemixObject object
}
\description{
In nonlinear mixed effect models, different types of predictions may be obtained, including individual predictions and population predictions
}
\keyword{methods}
