% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_generics.R, R/SaemixData.R,
%   R/SaemixModel.R, R/SaemixObject.R
\docType{methods}
\name{summary-methods}
\alias{summary-methods}
\alias{summary,ANY-method}
\alias{summary}
\alias{summary,SaemixData}
\alias{summary,SaemixData-method}
\alias{summary,SaemixModel-method}
\alias{summary,SaemixObject-method}
\title{Methods for Function summary}
\usage{
\S4method{summary}{SaemixData}(object, print = TRUE, ...)

\S4method{summary}{SaemixModel}(object, print = TRUE, ...)

\S4method{summary}{SaemixObject}(object, print = TRUE, ...)
}
\arguments{
\item{object}{an object of class SaemixData}

\item{print}{a boolean controlling whether to print the output or return it silently}

\item{...}{additional arguments (ignored)}
}
\value{
a list with a number of elements extracted from the dataset
\describe{
\item{N}{ number of subjects}
\item{nobs}{ the total number of observations} 
\item{nind.obs}{a vector giving the number of observations for each subject}
\item{id}{subject ID; x: predictors; y: response, and, if present in the data, covariates: the covariates (as many lines as observations) and ind.covariates: the individual covariates (one}
}
}
\description{
Methods for function \code{summary}

summary method for class SaemixData
}
\section{Methods}{
 \describe{

\item{list("signature(x = \"ANY\")")}{ default summary function ?}

\item{list("signature(x = \"SaemixData\")")}{ summary of the data }

\item{list("signature(x = \"SaemixModel\")")}{ summary of the model }

\item{list("signature(x = \"SaemixObject\")")}{ summary of an SaemixObject}

}
}

\keyword{methods}
